% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{diva_setup}
\alias{diva_setup}
\title{Setup divandr}
\usage{
diva_setup(pkg_check = TRUE, ...)
}
\arguments{
\item{pkg_check}{logical, check for DIVAnd.jl package and install if necessary}

\item{...}{Parameters are passed down to JuliaCall::julia_setup}
}
\description{
This function initializes Julia and the DIVAnd.jl package.
The first time will be long since it includes precompilation.
Additionally, this will install Julia and the required packages
if they are missing.
}
\examples{

\dontrun{ ## divandr_setup() is time-consuming and requires Julia+DIVAnd.jl
library(divandr)

diva_setup()

}

}
